requirements.txt
setuptools>=65.5.0
wheel
python-dotenv
pandas
numpy>=1.21.0
pymongo[srv]==4.3.3
certifi
scikit-learn==1.6.1
dill
pyaml
PyYAML
joblib
scipy==1.9.3
threadpoolctl
xgboost==1.7.6
category_encoders
fastapi==0.100.0
pydantic==1.10.9
python-dateutil
pytz
tzdata
statsmodels==0.13.5
patsy
typing-extensions
typing-inspection
annotated-types
pydantic-core
anyio
starlette
sniffio
exceptiongroup
idna
uvicorn==0.22.0
jinja2
python-multipart
